Move CRFModel RDD predict method to extension class
The CRFModel class provides a predict method that takes an array,
as well as one that takes an RDD. This in principle allows it to
be used without Spark, but because the class still contains the
method that depnds on the RDD type, it must be compiled against
the Spark libs even when that method is not used.

This commit moves the RDD overload of the predict method to an
implicit class; now, in order to use this method, you must
explicitly request it by importing nlp.Implicits._. If you do not
use this method, you do not need the Spark libs to use CRFModel.